requirements.txt
flask
flask-restful
flasgger
pyjwt
python-simplexml
python-dotenv
flask-cors
email_validator

# flask-sqlalchemy
# pymysql
# factory_boy
# alembic
# uwsgi
# pytest-cov